requirements.txt
flask==2.1.1
werkzeug==2.1.1
mysql-connector-python==8.0.27